+++ ksa_compliance/ksa_compliance/doctype/zatca_business_settings/zatca_business_settings.py
@@ -108,6 +108,7 @@ class ZATCABusinessSettings(Document):
             'Zero rated goods || Private healthcare to citizen',
             'Zero rated goods || Supply of qualified military goods',
         ]
+
     # end: auto-generated types
 
     def after_insert(self):
@@ -141,7 +142,8 @@ class ZATCABusinessSettings(Document):
     @property
     def has_production_csid(self) -> bool:
         return (
-            bool(self.production_security_token) and bool(self.production_secret) and bool(self.production_request_id)
+                bool(self.production_security_token) and bool(self.production_secret) and bool(
+            self.production_request_id)
         )
 
     @property
@@ -280,7 +282,7 @@ class ZATCABusinessSettings(Document):
 
     @staticmethod
     def for_invoice(
-        invoice_id: str, doctype: Literal['Sales Invoice', 'POS Invoice']
+            invoice_id: str, doctype: Literal['Sales Invoice', 'POS Invoice']
     ) -> Optional['ZATCABusinessSettings']:
         company_id = frappe.db.get_value(doctype, invoice_id, ['company'])
         if not company_id:
@@ -289,10 +291,17 @@ class ZATCABusinessSettings(Document):
         return ZATCABusinessSettings.for_company(company_id)
 
     @staticmethod
-    def for_company(company_id: str) -> Optional['ZATCABusinessSettings']:
+    def for_company(company_id: str, include_revoked=False) -> Optional['ZATCABusinessSettings']:
         business_settings_id = frappe.db.get_value(
             'ZATCA Business Settings', filters={'company': company_id, 'status': 'Active'}
         )
+        if not business_settings_id and include_revoked:
+            # frappe.db.exists doesn't order, so it could return the oldest revoked settings. We want the most
+            # recent revoked settings instead
+            business_settings_id = frappe.db.get_value('ZATCA Business Settings',
+                                                       {'company': company_id, 'status': 'Revoked'},
+                                                       ignore=True, order_by='modified desc')
+
         if not business_settings_id:
             return None